requirements.txt
# ===== CORE FLASK FRAMEWORK =====
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# ===== AZURE SERVICES =====
# Azure OpenAI and Cognitive Services
openai==1.13.3
azure-identity==1.15.0
azure-core==1.29.5

# Azure Cosmos DB
azure-cosmos==4.9.0

# Azure Storage (optional)
azure-storage-blob==12.19.0

# ===== ASYNC AND HTTP =====
aiohttp==3.9.3
asyncio==3.4.3
httpx==0.27.0

# ===== ENVIRONMENT AND CONFIG =====
python-dotenv==1.0.1
pydantic==2.5.3

# ===== DATA PROCESSING =====
# For document processing
pandas==2.1.4
numpy==1.24.4

# For text processing
nltk==3.8.1
spacy==3.7.2

# ===== FILE HANDLING =====
# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.0

# Word documents
python-docx==1.1.0

# Excel files
openpyxl==3.1.2

# ===== LOGGING AND MONITORING =====
structlog==23.2.0

# ===== UTILITIES =====
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# ===== TYPING AND VALIDATION =====
typing-extensions==4.9.0
annotated-types==0.6.0